+++ Typescript/src/baker.ts
@@ -1,0 +1,32 @@
+import { Clock } from "./clock";
+import { CalendarDate, Size, WeekDay } from "./index";
+import { Worker } from "./worker";
+
+export class Baker extends Worker {
+  static WORK_DAYS: WeekDay[] = [
+    "Monday",
+    "Tuesday",
+    "Wednesday",
+    "Thursday",
+    "Friday",
+  ];
+
+  constructor() {
+    super(Baker.WORK_DAYS);
+  }
+
+  public bake(
+    startDay: CalendarDate,
+    size: Size,
+    isMorningOrder?: boolean
+  ): CalendarDate {
+    let fullDaysOfWorkRequired = size === "big" ? 3 : 2;
+    if (!isMorningOrder) fullDaysOfWorkRequired++;
+
+    const daysToDoWork = this.calculateDaysToDoWork(
+      startDay.day,
+      fullDaysOfWorkRequired
+    );
+    return new Clock(startDay).add(daysToDoWork).toCalendar();
+  }
+}
